requirements/base.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
cryptography==37.0.2
dependency-injector==4.39.1
ecdsa==0.17.0
execnet==1.9.0
fastapi==0.93.0
greenlet==1.1.2
h11==0.13.0
httpcore==0.16.3
httptools==0.4.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
isort==5.10.1
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-forked==1.4.0
pytest-mock==3.9.0
pytest-xdist==2.5.0
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.16.0
watchfiles==0.17.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
pandas
joblib
scikit-learn
boto3